Code split edit dialog
